+++ packages/runtime/src/lib/txn.ts
@@ -10,6 +10,7 @@ import { Interpreter } from "..";
 import { RUNTIME_ERRORS } from "../errors/errors-list";
 import { RuntimeError } from "../errors/runtime-errors";
 import { Op } from "../interpreter/opcode";
+import { ITxn, ITxna } from "../interpreter/opcode-list";
 import {
 	ALGORAND_MIN_TX_FEE,
 	TransactionTypeEnum,
@@ -17,7 +18,15 @@ import {
 	TxnFields,
 	ZERO_ADDRESS_STR,
 } from "../lib/constants";
-import { Context, EncTx, RuntimeAccountI, StackElem, TxField, TxnType } from "../types";
+import {
+	Context,
+	EncTx,
+	ExecutionMode,
+	RuntimeAccountI,
+	StackElem,
+	TxField,
+	TxnType,
+} from "../types";
 import { convertToString } from "./parsing";
 
 export const assetTxnFields = new Set([
@@ -481,4 +490,46 @@ export function calculateFeeCredit(groupTx: EncTx[]): CreditFeeType {
 		collectedFee,
 		requiredFee,
 	};
+}
+
+/**
+ * Retunrs field f of the last inner transaction
+ * @param op ITxna or ITxn opcode
+ * @returns result
+ */
+export function executeITxn(op: ITxna | ITxn): StackElem {
+	const groupTx = op.interpreter.innerTxnGroups[op.interpreter.innerTxnGroups.length - 1];
+	const tx = groupTx[groupTx.length - 1];
+	let result: StackElem;
+	switch (op.field) {
+		case "Logs": {
+			const txReceipt = op.interpreter.runtime.ctx.state.txReceipts.get(tx.txID);
+			const logs: Uint8Array[] = txReceipt?.logs ?? [];
+			op.checkIndexBound(op.idx, logs, op.line);
+			result = logs[op.idx];
+			break;
+		}
+		case "NumLogs": {
+			const txReceipt = op.interpreter.runtime.ctx.state.txReceipts.get(tx.txID);
+			const logs: Uint8Array[] = txReceipt?.logs ?? [];
+			result = BigInt(logs.length);
+			break;
+		}
+		case "CreatedAssetID": {
+			result = BigInt(op.interpreter.runtime.ctx.createdAssetID);
+			break;
+		}
+		case "CreatedApplicationID": {
+			result = 0n; // can we create an app in inner-tx?
+			break;
+		}
+		default: {
+			result = txnSpecByField(op.field, tx, [tx], op.interpreter.tealVersion);
+			if (result === undefined || Object(result).length === 0) {
+				result = txAppArg(op.field, tx, op.idx, op, op.interpreter, op.line);
+				break;
+			}
+		}
+	}
+	return result;
 }